requirements.txt
Django
djangorestframework
pandas
numpy
scikit-learn
scipy
tqdm
requests
Pillow
